# Conflicts:
#	yais.sh

+++ chartman.sh
@@ -5,7 +5,7 @@
 dockerImage=${CHARTMAN_DOCKER_IMAGE:-docker-registry.ds.mhie.com/chartman}
 versionRequest=${CHARTMAN_VERSION:-3.x}
 cacheTtl=${CHARTMAN_CACHE_TTL:-300}
-minimumRunVersion="3.5.0"
+minimumRunVersion="3.11.0"
 cacheFilePath="/tmp/chartman/request-$versionRequest"
 latestFilePath="/tmp/chartman/latest"
 tracesFilePath="/dev/null"
@@ -41,6 +41,7 @@ check_file_modified_within_seconds() {
   fi
 }
 
+output=""
 read_run_parameters() {
   runParameters="$1"
 
@@ -54,12 +55,17 @@ fetch_run_parameters() {
 
   trace "Fetching run parameters with version $version"
 
-  fetchParameters=`docker run --rm -e CHARTMAN_DOCKER_REGISTRY_TOKEN=$CHARTMAN_DOCKER_REGISTRY_TOKEN -v $HOME/.chartman:/root/.chartman -v $HOME/.docker:/root/.docker $dockerImage:$version internal get-run-parameters --image-url $dockerImage $versionRequest 2>>"$tracesFilePath"`
+  fetchParameters=`docker run --rm -e CHARTMAN_DOCKER_REGISTRY_TOKEN=$CHARTMAN_DOCKER_REGISTRY_TOKEN -v $HOME/.chartman:/root/.chartman -v $HOME/.docker:/root/.docker $dockerImage:$version internal get-run-parameters --image-url $dockerImage $versionRequest 2>&1`
   runSucceed=$?
 
   if [ "$runSucceed" -eq 0 ]; then
     runParametersOutput="$fetchParameters"
     read_run_parameters "$runParametersOutput"
+  else
+    if [ "$CHARTMAN_TRACE_ENABLED" = "1" ]; then
+        echo "$fetchParameters" >> $tracesFilePath
+    fi
+    output=$fetchParameters
   fi
 }
 
@@ -120,5 +126,11 @@ resolvedArgs=$(eval echo \"$runDockerArgs\")
 
 trace "Resolved args: $resolvedArgs"
 trace "docker run --rm $dockerArgs $resolvedArgs -w $PWD $dockerImage:$requestedVersion"
+
+
+if [ "$requestedVersion" == "" ]; then
+  echo "$output"
+  exit 1
+fi
 
 docker run --rm $dockerArgs $resolvedArgs -w $PWD $dockerImage:$requestedVersion "$@"
